refactor(land): move registry files to repo directory
