+++ src-devtool/impl/viewer/internals/LandCacheViewer.cc
@@ -1,7 +1,8 @@
 #include "LandCacheViewer.h"
 
 #include "pland/PLand.h"
-#include "pland/land/LandRegistry.h"
+#include "pland/land/repo/LandRegistry.h"
+
 
 #include "mc/platform/UUID.h"
