+++ src/pland/PLand.cpp
@@ -21,10 +21,11 @@
 #include "pland/hooks/EventListener.h"
 #include "pland/infra/Config.h"
 #include "pland/infra/SafeTeleport.h"
-#include "pland/land/LandRegistry.h"
 #include "pland/land/LandScheduler.h"
+#include "pland/land/repo/LandRegistry.h"
 #include "pland/selector/SelectorManager.h"
 #include "pland/service/ServiceLocator.h"
+
 
 #ifdef LD_DEVTOOL
 #include "DevToolApp.h"
