+++ src/pland/gui/LandBuyGUI.cc
@@ -4,6 +4,7 @@
 
 #include "mc/world/actor/player/Player.h"
 
+#include "pland/land/repo/StorageError.h"
 #include "pland/Global.h"
 #include "pland/PLand.h"
 #include "pland/aabb/LandAABB.h"
@@ -14,9 +15,8 @@
 #include "pland/land/Land.h"
 #include "pland/land/LandCreateValidator.h"
 #include "pland/land/LandEvent.h"
-#include "pland/land/LandRegistry.h"
 #include "pland/land/LandResizeSettlement.h"
-#include "pland/land/StorageError.h"
+#include "pland/land/repo/LandRegistry.h"
 #include "pland/selector/DefaultSelector.h"
 #include "pland/selector/LandResizeSelector.h"
 #include "pland/selector/SelectorManager.h"
@@ -26,6 +26,7 @@
 #include "pland/service/ServiceLocator.h"
 #include "pland/utils/FeedbackUtils.h"
 #include "pland/utils/McUtils.h"
+
 
 #include <climits>
 #include <string>
