+++ src/pland/gui/LandManagerGUI.cc
@@ -11,6 +11,8 @@
 #include "mc/deps/ecs/WeakEntityRef.h"
 #include "mc/world/actor/player/Player.h"
 
+#include "pland/land/repo/LandContext.h"
+#include "pland/land/repo/StorageError.h"
 #include "pland/Global.h"
 #include "pland/PLand.h"
 #include "pland/aabb/LandAABB.h"
@@ -22,11 +24,9 @@
 #include "pland/gui/form/BackSimpleForm.h"
 #include "pland/infra/Config.h"
 #include "pland/land/Land.h"
-#include "pland/land/LandContext.h"
 #include "pland/land/LandCreateValidator.h"
 #include "pland/land/LandEvent.h"
-#include "pland/land/LandRegistry.h"
-#include "pland/land/StorageError.h"
+#include "pland/land/repo/LandRegistry.h"
 #include "pland/selector/LandResizeSelector.h"
 #include "pland/selector/SelectorManager.h"
 #include "pland/service/LandHierarchyService.h"
@@ -35,6 +35,7 @@
 #include "pland/service/ServiceLocator.h"
 #include "pland/utils/FeedbackUtils.h"
 #include "pland/utils/McUtils.h"
+
 
 #include <cstdint>
 #include <stack>
