+++ src/pland/gui/LandTeleportGUI.cc
@@ -6,8 +6,9 @@
 #include "pland/gui/form/BackSimpleForm.h"
 #include "pland/infra/SafeTeleport.h"
 #include "pland/land/Land.h"
-#include "pland/land/LandRegistry.h"
+#include "pland/land/repo/LandRegistry.h"
 #include "pland/utils/McUtils.h"
+
 
 
 namespace land {
