+++ src/pland/gui/common/EditLandPermTableUtilGUI.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "pland/land/LandContext.h"
+#include "pland/land/repo/LandContext.h"
 #include <functional>
 
 class Player;
