+++ src/pland/service/LandManagementService.cc
@@ -10,12 +10,13 @@
 #include "pland/events/player/PlayerRequestCreateLandEvent.h"
 #include "pland/infra/Config.h"
 #include "pland/land/LandCreateValidator.h"
-#include "pland/land/LandRegistry.h"
 #include "pland/land/LandResizeSettlement.h"
+#include "pland/land/repo/LandRegistry.h"
 #include "pland/selector/DefaultSelector.h"
 #include "pland/selector/LandResizeSelector.h"
 #include "pland/selector/SelectorManager.h"
 #include "pland/selector/SubLandSelector.h"
+
 
 #include <ll/api/event/EventBus.h>
